feature: implemented server state
