--- src/utils/customState.ts
@@ -1,13 +1,0 @@
-export function useCustomState(defaultValue: any) {
-  let value = defaultValue
-
-  function getValue() {
-    return value
-  }
-
-  function setValue(newValue: any) {
-    value = newValue
-  }
-
-  return [getValue, setValue];
-}
